refactor(services): split cell/space_manager.rs into submodules
Split the 1,420-line space_manager.rs into a space_manager/ directory:
- mod.rs: types + struct definition + tests
- xml.rs: XML loading
- lifecycle.rs: space create/destroy
- entities.rs: entity create/destroy/connect/disconnect/update
- aoi.rs: AoI computation
- spawn.rs: NPC spawning
- queries.rs: read-only accessors
- spatial.rs: line-of-sight, pathfinding, position validation

Public API unchanged. No behavior changes.
